Rename interactive plotting module for clarity
