+++ src/poliastro/plotting/__init__.py
@@ -1,4 +1,4 @@
-from poliastro.plotting.core import OrbitPlotter2D, OrbitPlotter3D
+from poliastro.plotting.interactive import OrbitPlotter2D, OrbitPlotter3D
 from poliastro.plotting.static import StaticOrbitPlotter
 
 __all__ = ["OrbitPlotter2D", "OrbitPlotter3D", "StaticOrbitPlotter"]
